+++ experimental/tools/unit-generators/vhdl/generators/handshake/muli.py
@@ -1,4 +1,4 @@
-from generators.support.oehb import generate_oehb
+from generators.handshake.oehb import generate_oehb
 from generators.support.join import generate_join
 from generators.support.delay_buffer import generate_delay_buffer
